[XrdClHttp] Omit Range for full-object prefetch (off_t max)
Full-download sets m_prefetch_size to off_t max (INT64_MAX), but
CurlReadOp only skipped Range for UINT64_MAX, so a huge Range was
sent and size became known via Content-Range. Treat INT64_MAX as
the whole-object sentinel instead.

+++ src/XrdClHttp/XrdClHttpOpRead.cc
@@ -123,9 +123,9 @@ CurlReadOp::Setup(CURL *curl, CurlWorker &worker)
     else if (m_op.second >= 128*1024) {
         curl_easy_setopt(curl, CURLOPT_BUFFERSIZE, 32*1024);
     }
-    // If the requested read size is UINT64_MAX, it means read the entire object;
+    // If the requested read size is INT64_MAX, it means read the entire object;
     // in this case, we do not set the Range header.
-    if (m_op.second != UINT64_MAX) {
+    if (m_op.second != static_cast<uint64_t>(INT64_MAX)) {
         auto range_req = "bytes=" + std::to_string(m_op.first) + "-" + std::to_string(m_op.first + m_op.second - 1);
         m_headers_list.emplace_back("Range", range_req);
     }
